chore!: Move CRT dependency from SmithyHTTPAPI to ClientRuntime (#1003)
